add fancy alert box

+++ src/pages/Dashboard/ServiceList/ServiceList.js
@@ -1,11 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import './ServiceList.css';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faEdit, faTrashAlt } from '@fortawesome/free-solid-svg-icons';
 import Sidebar from '../Sidebar/Sidebar';
 import { DataGrid } from "@material-ui/data-grid";
 import DeleteOutlineIcon from "@material-ui/icons/DeleteOutline";
 import { Link } from "react-router-dom";
+import Swal from 'sweetalert2';
 
 const ServiceList = () => {
     const [services, setServices] = useState([]);
@@ -23,18 +22,41 @@ const ServiceList = () => {
         setServices(services.filter((item) => item._id !== ids))
     }, [ids])
 
+
     // Delete service onClick
     const deleteService = (id) => {
-        setIds(id)
-        console.log(id, 'clicked')
-        fetch(`http://localhost:5000/deleteService/${id}`, {
-            method: 'DELETE',
+
+        // Fancy pop up
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                Swal.fire(
+                    'Deleted!',
+                    'Your file has been deleted.',
+                    'success'
+                );
+
+                setIds(id)
+                console.log(id, 'clicked')
+                fetch(`http://localhost:5000/deleteService/${id}`, {
+                    method: 'DELETE',
+                })
+                    .then(res => res.json())
+                    .then(result => { console.log(result) })
+            }
         })
-            .then(res => res.json())
-            .then(result => {
-                console.log(result)
-            })
+
     }
+
+
+
 
     const columns = [
         { field: "_id", hide: true },
@@ -86,23 +108,6 @@ const ServiceList = () => {
         },
     ];
 
-    // const editService = (id) => {
-    //     const editData = {
-
-    //     }
-    //     fetch(`https://e-sheba.herokuapp.com/editService/${id}`, {
-    //         method: 'PATCH',
-    //         headers: {
-    //             'content-type': 'application/json'
-    //         },
-    //         body: JSON.stringify(editData)
-    //     })
-    //         .then(res => res.json())
-    //         .then(data => {
-    //             console.log('updated')
-    //         })
-    // }
-
     return (
         <section className="serviceList">
             <div className="serviceListLeft">
@@ -125,44 +130,3 @@ const ServiceList = () => {
 };
 
 export default ServiceList;
-
-{/* onClick={() => editService(service._id)} */ }
-{/*  onClick={() => deleteService(service._id)} */ }
-
-
-
-//     <section className="serviceList">
-        //         <div className="serviceListLeft">
-        //             <Sidebar></Sidebar>
-        //         </div>
-        //         <div className="serviceListRight">
-        //             <table class="table table-striped">
-        //                 <thead>
-        //                     <tr>
-        //                         <th scope="col">Service Name</th>
-        //                         <th scope="col">Image</th>
-        //                         <th scope="col">Price</th>
-        //                         <th scope="col">Available</th>
-        //                         <th scope="col">Action</th>
-        //                     </tr>
-        //                 </thead>
-        //                 { 
-        //                     services.map(service =>
-        //                         <tbody>
-        //                             <tr>
-        //                                 <td>{service.serviceName}</td>
-        //                                 <td><img className="serviceImageDashboard" src={service.image} alt="" /></td>
-        //                                 <td>$ {service.price}</td>
-        //                                 <td>{service.isAvaiable}</td>
-        //                                 <td>
-        //                                     <div className="d-flex align-items-center">
-        //                                         <button className="btn btn-success me-2"><FontAwesomeIcon icon={faEdit} /></button>
-        //                                         <button className="btn btn-danger"><FontAwesomeIcon icon={faTrashAlt} /></button>
-        //                                     </div>
-        //                                 </td>
-        //                             </tr>
-        //                         </tbody>)
-        //                 }
-        //             </table>
-        //         </div>
-        // </section>
